+++ src/Export/BookExport.php
@@ -6,8 +6,9 @@
  * @author     Didier Corbière <contact@atoll-digital-library.org>
  */
 
-namespace Marsender\EPubLoader;
+namespace Marsender\EPubLoader\Export;
 
+use Marsender\EPubLoader\Metadata\BookInfos;
 use Exception;
 
 class BookExport